get rids of the whole error_message module since it has been integrated into the last version of deserr
