+++ meilisearch-types/src/keys.rs
@@ -11,8 +11,7 @@ use time::macros::{format_description, time};
 use time::{Date, OffsetDateTime, PrimitiveDateTime};
 use uuid::Uuid;
 
-use crate::deserr::error_messages::immutable_field_error;
-use crate::deserr::{DeserrError, DeserrJsonError};
+use crate::deserr::{immutable_field_error, DeserrError, DeserrJsonError};
 use crate::error::deserr_codes::*;
 use crate::error::{unwrap_any, Code, ErrorCode, ParseOffsetDateTimeError};
 use crate::index_uid_pattern::{IndexUidPattern, IndexUidPatternFormatError};
