+++ meilisearch/src/routes/indexes/mod.rs
@@ -5,9 +5,8 @@ use actix_web::{web, HttpRequest, HttpResponse};
 use deserr::{DeserializeError, Deserr, ValuePointerRef};
 use index_scheduler::IndexScheduler;
 use log::debug;
-use meilisearch_types::deserr::error_messages::immutable_field_error;
 use meilisearch_types::deserr::query_params::Param;
-use meilisearch_types::deserr::{DeserrJsonError, DeserrQueryParamError};
+use meilisearch_types::deserr::{immutable_field_error, DeserrJsonError, DeserrQueryParamError};
 use meilisearch_types::error::deserr_codes::*;
 use meilisearch_types::error::{unwrap_any, Code, ResponseError};
 use meilisearch_types::index_uid::IndexUid;
